First draft of client connection
