+++ scripting_client/planetaryimager/network/__init__.py
@@ -1,4 +1,6 @@
 from .network_packet import NetworkPacket
 from .protocol import Protocol
 from .hello import Hello
+from .client import Client
+
