+++ scripting_client/planetaryimager/planetaryimager_client.py
@@ -1,6 +1,12 @@
+from .network import Client, Hello
 
 class PlanetaryImagerClient:
     def __init__(self, address, port=19232):
-        self.address = address
-        self.port = port
+        self.client = Client(address, port)
 
+    def connect(self):
+        self.client.connect()
+        rec = self.client.send_receive(Hello.send_packet())
+        return rec
+
+
